Article 도메인에 대한 서비스, 리포지터리 도입

+++ src/main/java/com/ll/exam/article/dto/ArticleDto.java
@@ -1,11 +1,11 @@
 package com.ll.exam.article.dto;
 
 public class ArticleDto {
-    private int id;
+    private long id;
     private String title;
     private String body;
 
-    public ArticleDto(int id, String title, String body) {
+    public ArticleDto(long id, String title, String body) {
         this.id = id;
         this.title = title;
         this.body = body;
@@ -20,7 +20,7 @@ public class ArticleDto {
                 '}';
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
